Time: 278 ms (43.32%), Space: 67.7 MB (93.74%) - LeetHub

+++ Pandas/0596-classes-with-at-least-5-students/0596-classes-with-at-least-5-students.py
@@ -3,5 +3,5 @@ import numpy as np
 
 def find_classes(courses: pd.DataFrame) -> pd.DataFrame:
     courses = courses.groupby('class')['student'].size().reset_index()
-    courses = courses[courses['student'].astype('float64') >= 5]['class']
+    courses = courses[courses['student'] >= 5]['class']
     return pd.DataFrame(courses)
